fix BUILD.gn tests
remove session_connection.h
remove vsync_waitr_unittests
