+++ shell/platform/fuchsia/flutter/tests/default_session_connection_unittests.cc
@@ -17,6 +17,16 @@ using namespace flutter_runner;
 
 namespace flutter_runner_test {
 
+static fml::TimePoint TimePointFromInt(int i) {
+  return fml::TimePoint::FromEpochDelta(fml::TimeDelta::FromNanoseconds(i));
+}
+static fml::TimeDelta TimeDeltaFromInt(int i) {
+  return fml::TimeDelta::FromNanoseconds(i);
+}
+static int TimePointToInt(fml::TimePoint time) {
+  return time.ToEpochDelta().ToNanoseconds();
+}
+
 class DefaultSessionConnectionTest : public ::testing::Test {
  public:
   void SetUp() override {
@@ -68,7 +78,7 @@ TEST_F(DefaultSessionConnectionTest, SimplePresentTest) {
 
   flutter_runner::DefaultSessionConnection session_connection(
       "debug label", std::move(session_), on_session_error_callback,
-      on_frame_presented_callback, 3);
+      on_frame_presented_callback, 3, fml::TimeDelta::FromSeconds(0));
 
   for (int i = 0; i < 200; ++i) {
     session_connection.Present();
@@ -90,7 +100,7 @@ TEST_F(DefaultSessionConnectionTest, BatchedPresentTest) {
 
   flutter_runner::DefaultSessionConnection session_connection(
       "debug label", std::move(session_), on_session_error_callback,
-      on_frame_presented_callback, 3);
+      on_frame_presented_callback, 3, fml::TimeDelta::FromSeconds(0));
 
   for (int i = 0; i < 200; ++i) {
     session_connection.Present();
@@ -100,16 +110,6 @@ TEST_F(DefaultSessionConnectionTest, BatchedPresentTest) {
   }
 
   EXPECT_GT(num_presents_handled, 0u);
-}
-
-static fml::TimePoint TimePointFromInt(int i) {
-  return fml::TimePoint::FromEpochDelta(fml::TimeDelta::FromNanoseconds(i));
-}
-static fml::TimeDelta TimeDeltaFromInt(int i) {
-  return fml::TimeDelta::FromNanoseconds(i);
-}
-static int TimePointToInt(fml::TimePoint time) {
-  return time.ToEpochDelta().ToNanoseconds();
 }
 
 // The first set of tests has an empty |future_presentation_infos| passed in.
@@ -468,5 +468,219 @@ TEST(CalculateNextLatchPointTest, SteadyState_FuzzyLatchPointsAfterTarget) {
             TimePointToInt(present_requested_time + flutter_frame_build_time));
   EXPECT_GE(TimePointToInt(calculated_latch_point),
             TimePointToInt(last_latch_point_targeted));
+}
+
+TEST(SnapToNextPhaseTest, SnapOverlapsWithNow) {
+  const auto now = fml::TimePoint::Now();
+  const auto last_presentation_time = now - fml::TimeDelta::FromNanoseconds(10);
+  const auto delta = fml::TimeDelta::FromNanoseconds(10);
+  const auto next_vsync =
+      flutter_runner::DefaultSessionConnection::SnapToNextPhase(
+          now, last_presentation_time, delta);
+
+  EXPECT_EQ(now + delta, next_vsync);
+}
+
+TEST(SnapToNextPhaseTest, SnapAfterNow) {
+  const auto now = fml::TimePoint::Now();
+  const auto last_presentation_time = now - fml::TimeDelta::FromNanoseconds(9);
+  const auto delta = fml::TimeDelta::FromNanoseconds(10);
+  const auto next_vsync =
+      flutter_runner::DefaultSessionConnection::SnapToNextPhase(
+          now, last_presentation_time, delta);
+
+  // math here: 10 - 9 = 1
+  EXPECT_EQ(now + fml::TimeDelta::FromNanoseconds(1), next_vsync);
+}
+
+TEST(SnapToNextPhaseTest, SnapAfterNowMultiJump) {
+  const auto now = fml::TimePoint::Now();
+  const auto last_presentation_time = now - fml::TimeDelta::FromNanoseconds(34);
+  const auto delta = fml::TimeDelta::FromNanoseconds(10);
+  const auto next_vsync =
+      flutter_runner::DefaultSessionConnection::SnapToNextPhase(
+          now, last_presentation_time, delta);
+
+  // zeroes: -34, -24, -14, -4, 6, ...
+  EXPECT_EQ(now + fml::TimeDelta::FromNanoseconds(6), next_vsync);
+}
+
+TEST(SnapToNextPhaseTest, SnapAfterNowMultiJumpAccountForCeils) {
+  const auto now = fml::TimePoint::Now();
+  const auto last_presentation_time = now - fml::TimeDelta::FromNanoseconds(20);
+  const auto delta = fml::TimeDelta::FromNanoseconds(16);
+  const auto next_vsync =
+      flutter_runner::DefaultSessionConnection::SnapToNextPhase(
+          now, last_presentation_time, delta);
+
+  // zeroes: -20, -4, 12, 28, ...
+  EXPECT_EQ(now + fml::TimeDelta::FromNanoseconds(12), next_vsync);
+}
+
+TEST(GetTargetTimesTest, ScheduleForNextVsync) {
+  const fml::TimeDelta vsync_offset = TimeDeltaFromInt(0);
+  const fml::TimeDelta vsync_interval = TimeDeltaFromInt(10);
+  const fml::TimePoint last_targetted_vsync = TimePointFromInt(10);
+  const fml::TimePoint now = TimePointFromInt(9);
+  const fml::TimePoint next_vsync = TimePointFromInt(10);
+
+  const auto target_times =
+      flutter_runner::DefaultSessionConnection::GetTargetTimes(
+          vsync_offset, vsync_interval, last_targetted_vsync, now, next_vsync);
+
+  EXPECT_EQ(TimePointToInt(target_times.frame_start), 10);
+  EXPECT_EQ(TimePointToInt(target_times.frame_target), 20);
+}
+
+TEST(GetTargetTimesTest, ScheduleForCurrentVsync_DueToOffset) {
+  const fml::TimeDelta vsync_offset = TimeDeltaFromInt(3);
+  const fml::TimeDelta vsync_interval = TimeDeltaFromInt(10);
+  const fml::TimePoint last_targetted_vsync = TimePointFromInt(0);
+  const fml::TimePoint now = TimePointFromInt(6);
+  const fml::TimePoint next_vsync = TimePointFromInt(10);
+
+  const auto target_times =
+      flutter_runner::DefaultSessionConnection::GetTargetTimes(
+          vsync_offset, vsync_interval, last_targetted_vsync, now, next_vsync);
+
+  EXPECT_EQ(TimePointToInt(target_times.frame_start), 7);
+  EXPECT_EQ(TimePointToInt(target_times.frame_target), 10);
+}
+
+TEST(GetTargetTimesTest, ScheduleForFollowingVsync_BecauseOfNow) {
+  const fml::TimeDelta vsync_offset = TimeDeltaFromInt(0);
+  const fml::TimeDelta vsync_interval = TimeDeltaFromInt(10);
+  const fml::TimePoint last_targetted_vsync = TimePointFromInt(10);
+  const fml::TimePoint now = TimePointFromInt(15);
+  const fml::TimePoint next_vsync = TimePointFromInt(10);
+
+  const auto target_times =
+      flutter_runner::DefaultSessionConnection::GetTargetTimes(
+          vsync_offset, vsync_interval, last_targetted_vsync, now, next_vsync);
+
+  EXPECT_EQ(TimePointToInt(target_times.frame_start), 20);
+  EXPECT_EQ(TimePointToInt(target_times.frame_target), 30);
+}
+
+TEST(GetTargetTimesTest, ScheduleForFollowingVsync_BecauseOfTargettedTime) {
+  const fml::TimeDelta vsync_offset = TimeDeltaFromInt(0);
+  const fml::TimeDelta vsync_interval = TimeDeltaFromInt(10);
+  const fml::TimePoint last_targetted_vsync = TimePointFromInt(20);
+  const fml::TimePoint now = TimePointFromInt(9);
+  const fml::TimePoint next_vsync = TimePointFromInt(10);
+
+  const auto target_times =
+      flutter_runner::DefaultSessionConnection::GetTargetTimes(
+          vsync_offset, vsync_interval, last_targetted_vsync, now, next_vsync);
+
+  EXPECT_EQ(TimePointToInt(target_times.frame_start), 20);
+  EXPECT_EQ(TimePointToInt(target_times.frame_target), 30);
+}
+
+TEST(GetTargetTimesTest, ScheduleForDistantVsync_BecauseOfTargettedTime) {
+  const fml::TimeDelta vsync_offset = TimeDeltaFromInt(0);
+  const fml::TimeDelta vsync_interval = TimeDeltaFromInt(10);
+  const fml::TimePoint last_targetted_vsync = TimePointFromInt(60);
+  const fml::TimePoint now = TimePointFromInt(9);
+  const fml::TimePoint next_vsync = TimePointFromInt(10);
+
+  const auto target_times =
+      flutter_runner::DefaultSessionConnection::GetTargetTimes(
+          vsync_offset, vsync_interval, last_targetted_vsync, now, next_vsync);
+
+  EXPECT_EQ(TimePointToInt(target_times.frame_start), 60);
+  EXPECT_EQ(TimePointToInt(target_times.frame_target), 70);
+}
+
+TEST(GetTargetTimesTest, ScheduleForFollowingVsync_WithSlightVsyncDrift) {
+  const fml::TimeDelta vsync_offset = TimeDeltaFromInt(0);
+  const fml::TimeDelta vsync_interval = TimeDeltaFromInt(10);
+
+  // Even though it appears as if the next vsync is at time 40, we should still
+  // present at time 50.
+  const fml::TimePoint last_targetted_vsync = TimePointFromInt(37);
+  const fml::TimePoint now = TimePointFromInt(9);
+  const fml::TimePoint next_vsync = TimePointFromInt(10);
+
+  const auto target_times =
+      flutter_runner::DefaultSessionConnection::GetTargetTimes(
+          vsync_offset, vsync_interval, last_targetted_vsync, now, next_vsync);
+
+  EXPECT_EQ(TimePointToInt(target_times.frame_start), 40);
+  EXPECT_EQ(TimePointToInt(target_times.frame_target), 50);
+}
+
+TEST(GetTargetTimesTest, ScheduleForAnOffsetFromVsync) {
+  const fml::TimeDelta vsync_offset = TimeDeltaFromInt(4);
+  const fml::TimeDelta vsync_interval = TimeDeltaFromInt(10);
+  const fml::TimePoint last_targetted_vsync = TimePointFromInt(10);
+  const fml::TimePoint now = TimePointFromInt(9);
+  const fml::TimePoint next_vsync = TimePointFromInt(10);
+
+  const auto target_times =
+      flutter_runner::DefaultSessionConnection::GetTargetTimes(
+          vsync_offset, vsync_interval, last_targetted_vsync, now, next_vsync);
+
+  EXPECT_EQ(TimePointToInt(target_times.frame_start), 16);
+  EXPECT_EQ(TimePointToInt(target_times.frame_target), 20);
+}
+
+TEST(GetTargetTimesTest, ScheduleMultipleTimes) {
+  const fml::TimeDelta vsync_offset = TimeDeltaFromInt(0);
+  const fml::TimeDelta vsync_interval = TimeDeltaFromInt(10);
+
+  fml::TimePoint last_targetted_vsync = TimePointFromInt(0);
+  fml::TimePoint now = TimePointFromInt(5);
+  fml::TimePoint next_vsync = TimePointFromInt(10);
+
+  for (int i = 0; i < 100; ++i) {
+    const auto target_times =
+        flutter_runner::DefaultSessionConnection::GetTargetTimes(
+            vsync_offset, vsync_interval, last_targetted_vsync, now,
+            next_vsync);
+
+    EXPECT_EQ(TimePointToInt(target_times.frame_start), 10 * (i + 1));
+    EXPECT_EQ(TimePointToInt(target_times.frame_target), 10 * (i + 2));
+
+    // Simulate the passage of time.
+    now = now + vsync_interval;
+    next_vsync = next_vsync + vsync_interval;
+    last_targetted_vsync = target_times.frame_target;
+  }
+}
+
+TEST(GetTargetTimesTest, ScheduleMultipleTimes_WithDelayedWakeups) {
+  // It is often the case that Flutter does not wake up when it intends to due
+  // to CPU contention. This test has DefaultSessionConnection wake up to
+  // schedule 0-4ns after when |now| should be - and we verify that the results
+  // should be the same as if there were no delay.
+  const fml::TimeDelta vsync_offset = TimeDeltaFromInt(0);
+  const fml::TimeDelta vsync_interval = TimeDeltaFromInt(10);
+
+  fml::TimePoint last_targetted_vsync = TimePointFromInt(0);
+  fml::TimePoint now = TimePointFromInt(5);
+  fml::TimePoint next_vsync = TimePointFromInt(10);
+
+  for (int i = 0; i < 100; ++i) {
+    const auto target_times =
+        flutter_runner::DefaultSessionConnection::GetTargetTimes(
+            vsync_offset, vsync_interval, last_targetted_vsync, now,
+            next_vsync);
+
+    const auto target_times_delay =
+        flutter_runner::DefaultSessionConnection::GetTargetTimes(
+            vsync_offset, vsync_interval, last_targetted_vsync,
+            now + TimeDeltaFromInt(i % 5), next_vsync);
+
+    EXPECT_EQ(TimePointToInt(target_times.frame_start),
+              TimePointToInt(target_times_delay.frame_start));
+    EXPECT_EQ(TimePointToInt(target_times.frame_target),
+              TimePointToInt(target_times_delay.frame_target));
+
+    // Simulate the passage of time.
+    now = now + vsync_interval;
+    next_vsync = next_vsync + vsync_interval;
+    last_targetted_vsync = target_times.frame_target;
+  }
 }
 }  // namespace flutter_runner_test
